pyproject.toml
[project]
name = "jiraiq-mcp"
version = "1.0.0"
description = "JiraIQ MCP Server - Jira analysis tools for Claude"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Pratap Yeragudipati", email = "pratap@example.com"}
]
license = {text = "MIT"}
keywords = ["jira", "claude", "mcp", "analysis", "project-management"]

dependencies = [
    "mcp>=1.0.0",
    "jira>=3.5.0",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/pyeragudipati/jiraiq-mcp"
Repository = "https://github.com/pyeragudipati/jiraiq-mcp"
Documentation = "https://github.com/pyeragudipati/jiraiq-mcp#readme"

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["jiraiq_mcp"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
